requirements.txt
mysql-connector-python~=8.0.20
Flask~=1.1.2
Werkzeug~=1.0.1